+++ utils/config/struct.go
@@ -17,20 +17,20 @@ type Config struct {
 	ServerAddr          string        `json:"HTTP"`
 	HTTP2               bool          `json:"HTTP2"`
 	HSTS                string        `json:"HSTS"`
+
 	TLSServerAddr       string        `json:"HTTPS"`
 	TLSCertFile         string        `json:"CERT"`
 	TLSKeyFile          string        `json:"KEY"`
+
 	LetsEncrypt         bool          `json:"LETSENCRYPT"`
 	LetsEncryptCacheDir string        `json:"LETSENCRYPT_CERT_DIR"`
 	LetsEncryptEmail    string        `json:"LETSENCRYPT_EMAIL"`
 	LetsEncryptHosts    string        `json:"LETSENCRYPT_HOSTS"`
+
 	APIPrefix           string        `json:"API_PREFIX"`
 	CORSOrigin          string        `json:"CORS_ORIGIN"`
 	ReadTimeout         time.Duration `json:"READ_TIMEOUT"`
 	WriteTimeout        time.Duration `json:"WRITE_TIMEOUT"`
-	UpdateInterval      time.Duration `json:"UPDATE_INTERVAL"`
-	TorUpdateInterval   time.Duration `json:"TOR_UPDATE_INTERVAL"`
-	RetryInterval       time.Duration `json:"RETRY_INTERVAL"`
 	UseXForwardedFor    bool          `json:"USE_X_FORWARDED_FOR"`
 	Silent              bool          `json:"SILENT"`
 	LogToStdout         bool          `json:"LOGTOSTDOUT"`
@@ -39,17 +39,32 @@ type Config struct {
 	RedisTimeout        time.Duration `json:"REDIS_TIMEOUT"`
 	MemcacheAddr        string        `json:"MEMCACHE"`
 	MemcacheTimeout     time.Duration `json:"MEMCACHE_TIMEOUT"`
+
 	RateLimitBackend    string        `json:"QUOTA_BACKEND"`
 	RateLimitInterval   time.Duration `json:"QUOTA_INTERVAL"`
 	RateLimitLimit      uint64        `json:"QUOTA_MAX"`
 	RateLimitBurst      int           `json:"QUOTA_BURST"`
-	UpdatesHost         string        `json:"UPDATES_HOST"`
-	LicenseKey          string        `json:"LICENSE_KEY"`
-	UserID              string        `json:"USER_ID"`
-	ProductID           string        `json:"PRODUCT_ID"`
-	ASNProductID        string        `json:"ASN_PRODUCT_ID"`
-	GuiDir              string        `json:"GUI"`
+
+	MMUserID            string        `json:"MM_USER_ID"`
+	MMLicenseKey        string        `json:"MM_LICENSE_KEY"`
+	MMProductID         string        `json:"MM_PRODUCT_ID"`
+	MMASNProductID      string        `json:"MM_ASN_PRODUCT_ID"`
+	MMUpdatesHost       string        `json:"MM_UPDATES_HOST"`
+	MMRetryInterval     time.Duration `json:"MM_RETRY_INTERVAL"`
+	MMUpdateInterval    time.Duration `json:"MM_UPDATE_INTERVAL"`
+
+	I2LToken			string		  `json:"I2L_TOKEN"`
+	I2LProductID		string		  `json:"I2L_PRODUCT_ID"`
+	I2LRetryInterval    time.Duration `json:"I2L_RETRY_INTERVAL"`
+	I2LUpdateInterval   time.Duration `json:"I2L_UPDATE_INTERVAL"`
+	I2LUpdatesHost      string        `json:"I2L_UPDATES_HOST"`
+
 	TorExitCheck      	string 		  `json:"TOR_EXIT"`
+	TorRetryInterval    time.Duration `json:"TOR_RETRY_INTERVAL"`
+	TorUpdateInterval   time.Duration `json:"TOR_UPDATE_INTERVAL"`
+	TorUpdatesHost      string        `json:"TOR_UPDATES_HOST"`
+
+	GuiDir              string        `json:"GUI"`
 
 	File     			string 		  `json:"-"`
 	RootDir     		string 		  `json:"-"`